Idiomatic math package
