+++ packages/math/types/pool/__tests__/weighted.spec.d.ts
@@ -1,0 +1,1 @@
+export {};
