+++ packages/math/types/pool/index.d.ts
@@ -1,0 +1,2 @@
+export * from "./estimates";
+export * from "./weighted";
