+++ packages/math/types/pool/weighted.d.ts
@@ -9,4 +9,3 @@ export declare function calcPoolOutGivenSingleIn(tokenBalanceIn: Dec, tokenWeigh
 export declare function calcSingleInGivenPoolOut(tokenBalanceIn: Dec, tokenWeightIn: Dec, poolSupply: Dec, totalWeight: Dec, poolAmountOut: Dec, swapFee: Dec): Dec;
 export declare function calcSingleOutGivenPoolIn(tokenBalanceOut: Dec, tokenWeightOut: Dec, poolSupply: Dec, totalWeight: Dec, poolAmountIn: Dec, swapFee: Dec): Dec;
 export declare function calcPoolInGivenSingleOut(tokenBalanceOut: Dec, tokenWeightOut: Dec, poolSupply: Dec, totalWeight: Dec, tokenAmountOut: Dec, swapFee: Dec): Dec;
-export declare function pow(base: Dec, exp: Dec): Dec;
