+++ packages/web/config/feature-flag.ts
@@ -39,11 +39,12 @@ export const ExternalIncentiveGaugeAllowList: {
   }[];
 } = {
   "1": [
-		{
-			gaugeId:'4301',
-			denom:'ibc/AA1C80225BCA7B32ED1FC6ABF8B8E899BEB48ECDB4B417FD69873C6D715F97E7'
-		}
-	],
+    {
+      gaugeId: "4301",
+      denom:
+        "ibc/AA1C80225BCA7B32ED1FC6ABF8B8E899BEB48ECDB4B417FD69873C6D715F97E7",
+    },
+  ],
   "3": [
     {
       gaugeId: "3178",
@@ -844,14 +845,16 @@ export const ExternalIncentiveGaugeAllowList: {
   "716": [
     {
       gaugeId: "4303",
-      denom: "ibc/AA1C80225BCA7B32ED1FC6ABF8B8E899BEB48ECDB4B417FD69873C6D715F97E7",
-    }
+      denom:
+        "ibc/AA1C80225BCA7B32ED1FC6ABF8B8E899BEB48ECDB4B417FD69873C6D715F97E7",
+    },
   ],
-  "717":[
+  "717": [
     {
-      gaugeId:"4302",
-      denom:"ibc/AA1C80225BCA7B32ED1FC6ABF8B8E899BEB48ECDB4B417FD69873C6D715F97E7",
-    }
+      gaugeId: "4302",
+      denom:
+        "ibc/AA1C80225BCA7B32ED1FC6ABF8B8E899BEB48ECDB4B417FD69873C6D715F97E7",
+    },
   ],
   "719": [
     {
